+++ examples/compile.rs
@@ -1,11 +1,11 @@
 // usage:  cargo run --release --example compile
 
 use std::collections::BTreeMap;
-use fasteval::{EmptyNamespace, Evaler};    // use this trait so we can call eval().
-use fasteval::Compiler;  // use this trait so we can call compile().
-fn main() -> Result<(), fasteval::Error> {
-    let parser = fasteval::Parser::new();
-    let mut slab = fasteval::Slab::new();
+use fasteval2::{EmptyNamespace, Evaler};    // use this trait so we can call eval().
+use fasteval2::Compiler;  // use this trait so we can call compile().
+fn main() -> Result<(), fasteval2::Error> {
+    let parser = fasteval2::Parser::new();
+    let mut slab = fasteval2::Slab::new();
     let mut map = BTreeMap::new();
 
     let expr_str = "sin(deg/360 * 2*pi())";
@@ -14,7 +14,7 @@ fn main() -> Result<(), fasteval::Error> {
         map.insert("deg".to_string(), deg as f64);
         // When working with compiled constant expressions, you can use the
         // eval_compiled*!() macros to save a function call:
-        let val = fasteval::eval_compiled!(compiled, &slab, &mut map);
+        let val = fasteval2::eval_compiled!(compiled, &slab, &mut map);
         eprintln!("sin({}°) = {}", deg, val);
     }