+++ examples/ns_cachedcallbacknamespace.rs
@@ -1,6 +1,6 @@
 // usage:  cargo run --release --example ns_cachedcallbacknamespace
 
-fn main() -> Result<(), fasteval::Error> {
+fn main() -> Result<(), fasteval2::Error> {
     let mut num_lookups = 0;
     let val = {
         let cb = |name:&str, _args:Vec<f64>| -> Option<f64> {
@@ -15,9 +15,9 @@ fn main() -> Result<(), fasteval::Error> {
                 _ => None,
             }
         };
-        let mut ns = fasteval::CachedCallbackNamespace::new(cb);
+        let mut ns = fasteval2::CachedCallbackNamespace::new(cb);
 
-        fasteval::ez_eval("x * (x + 1)", &mut ns)?
+        fasteval2::ez_eval("x * (x + 1)", &mut ns)?
     };
     assert_eq!(val, 6.0);
     assert_eq!(num_lookups, 1);  // Notice that only 1 lookup occurred.