+++ examples/ns_emptynamespace.rs
@@ -1,9 +1,9 @@
 // usage:  cargo run --release --example ns_emptynamespace
 
-fn main() -> Result<(), fasteval::Error> {
-    let mut ns = fasteval::EmptyNamespace;
+fn main() -> Result<(), fasteval2::Error> {
+    let mut ns = fasteval2::EmptyNamespace;
 
-    let val = fasteval::ez_eval("sin(pi()/2)", &mut ns)?;
+    let val = fasteval2::ez_eval("sin(pi()/2)", &mut ns)?;
     assert_eq!(val, 1.0);
 
     Ok(())