+++ examples/unsafe-vars.rs
@@ -1,8 +1,8 @@
 // usage:  cargo run --release --features unsafe-vars --example unsafe-vars
 
-use fasteval::EmptyNamespace;
+use fasteval2::EmptyNamespace;
 
-fn main() -> Result<(), fasteval::Error> {
+fn main() -> Result<(), fasteval2::Error> {
     #[cfg(not(feature="unsafe-vars"))]
     {
         panic!("You must enable the `unsafe-vars` feature to run this example:  cargo run --release --features unsafe-vars --example unsafe-vars");
@@ -12,11 +12,11 @@ fn main() -> Result<(), fasteval::Error> {
     // This is important so that `cargo test` can succeed.
     #[cfg(feature="unsafe-vars")]
     {
-        use fasteval::Evaler;    // use this trait so we can call eval().
-        use fasteval::Compiler;  // use this trait so we can call compile().
+        use fasteval2::Evaler;    // use this trait so we can call eval().
+        use fasteval2::Compiler;  // use this trait so we can call compile().
 
-        let parser = fasteval::Parser::new();
-        let mut slab = fasteval::Slab::new();
+        let parser = fasteval2::Parser::new();
+        let mut slab = fasteval2::Slab::new();
 
         // The Unsafe Variable will use a pointer to read this memory location:
         // You must make sure that this variable stays in-scope as long as the
@@ -35,7 +35,7 @@ fn main() -> Result<(), fasteval::Error> {
 
         for d in 0..360 {
             deg = d as f64;
-            let val = fasteval::eval_compiled!(compiled, &slab, &mut ns);
+            let val = fasteval2::eval_compiled!(compiled, &slab, &mut ns);
             eprintln!("sin({}°) = {}", deg, val);
         }