+++ tests/evalns.rs
@@ -1,8 +1,8 @@
-use fasteval::ez_eval;
+use fasteval2::ez_eval;
 
 #[test]
 fn empty() {
-    let mut ns = fasteval::EmptyNamespace;
+    let mut ns = fasteval2::EmptyNamespace;
 
     let val = ez_eval("1 + 1", &mut ns).unwrap();
     assert_eq!(val, 2.0);
@@ -11,7 +11,7 @@ fn empty() {
 #[test]
 fn str_to_f64() {
     {
-        let mut ns = fasteval::StringToF64Namespace::new();
+        let mut ns = fasteval2::StringToF64Namespace::new();
         ns.insert("a".to_string(), 1.11);
         ns.insert("b".to_string(), 2.22);
 
@@ -20,7 +20,7 @@ fn str_to_f64() {
     }
 
     {
-        let mut ns = fasteval::StrToF64Namespace::new();
+        let mut ns = fasteval2::StrToF64Namespace::new();
         ns.insert("a", 1.11);
         ns.insert("b", 2.22);
 
@@ -32,7 +32,7 @@ fn str_to_f64() {
 #[test]
 fn str_to_cb() {
     {
-        let mut ns = fasteval::StringToCallbackNamespace::new();
+        let mut ns = fasteval2::StringToCallbackNamespace::new();
         ns.insert("a".to_string(), Box::new(|args| args[0]));
         ns.insert("b".to_string(), Box::new(|args| args[0] * 2.0));
 
@@ -41,7 +41,7 @@ fn str_to_cb() {
     }
 
     {
-        let mut ns = fasteval::StrToCallbackNamespace::new();
+        let mut ns = fasteval2::StrToCallbackNamespace::new();
         ns.insert("a", Box::new(|args| args[0]));
         ns.insert("b", Box::new(|args| args[0] * 2.0));
 
@@ -52,8 +52,8 @@ fn str_to_cb() {
 
 #[test]
 fn layered_str_to_f64() {
-    let mut ns = fasteval::LayeredStringToF64Namespace::new();
-    let mut layer0 = fasteval::StringToF64Namespace::new();
+    let mut ns = fasteval2::LayeredStringToF64Namespace::new();
+    let mut layer0 = fasteval2::StringToF64Namespace::new();
     layer0.insert("a".to_string(), 1.11);
     layer0.insert("b".to_string(), 2.22);
     ns.push(layer0);
@@ -61,7 +61,7 @@ fn layered_str_to_f64() {
     let val = ez_eval("a + b + 1", &mut ns).unwrap();
     assert_eq!(val, 4.33);
 
-    let mut layer1 = fasteval::StringToF64Namespace::new();
+    let mut layer1 = fasteval2::StringToF64Namespace::new();
     layer1.insert("a".to_string(), 11.11);
     ns.push(layer1);
 
@@ -91,7 +91,7 @@ fn cb() {
 
 #[test]
 fn cached_cb() {
-    let mut ns = fasteval::CachedCallbackNamespace::new(|name:&str, args:Vec<f64>| {
+    let mut ns = fasteval2::CachedCallbackNamespace::new(|name:&str, args:Vec<f64>| {
         match name {
             "a" => { eprintln!("cached_cb: a: This should only be printed once."); Some(1.11) }
             "b" => Some(2.22),
@@ -112,7 +112,7 @@ fn cached_cb() {
 fn custom_vector_funcs() {
     let vecs_cell = std::cell::RefCell::new(Vec::<Vec<f64>>::new());
 
-    let mut ns = fasteval::StrToCallbackNamespace::new();
+    let mut ns = fasteval2::StrToCallbackNamespace::new();
 
     ns.insert("x", Box::new(|_args| 2.0));