+++ tests/ez.rs
@@ -1,4 +1,4 @@
-use fasteval::{ez_eval, Error};
+use fasteval2::{ez_eval, Error};
 
 use std::collections::BTreeMap;